requirements.txt
discord.py==2.3.2
typer==0.9.0
rich==13.7.0
requests==2.31.0
python-dotenv==1.0.0
plotext==5.2.8
termgraph==0.5.3 